reorganized things a tad.  and also added a route
